+++ forge/src/main/java/dev/aika/smsn/forge/mixin/MixinInfo.java
@@ -17,6 +17,7 @@ public class MixinInfo {
                 "dev.aika.smsn.forge.mixin.ipn.IPNInfoManagerMixin", "inventoryprofilesnext",
                 "dev.aika.smsn.forge.mixin.projecte.ThreadCheckUUIDMixin", "projecte",
                 "dev.aika.smsn.forge.mixin.quark.ThreadContributorListLoaderMixin", "quark",
+                "dev.aika.smsn.forge.mixin.quark.QButtonMixin", "quark",
                 "dev.aika.smsn.forge.mixin.titanium.RewardMixin", "titanium",
                 "dev.aika.smsn.forge.mixin.xaero.minimap.InternetMixin", "xaerominimap",
                 "dev.aika.smsn.forge.mixin.xaero.minimap.PatreonMixin", "xaerominimap",