+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+### Unreleased
+
+**Forge**
+
+- Fewer Quark Celebration
+
 ### 1.2.7
 
-**NeoForge**
+**Forge**
 
 - Add Mixins for [ProjectE](https://www.curseforge.com/minecraft/mc-mods/projecte)
 - Refactoring the configuration API
@@ -13,7 +19,7 @@
 
 ### 1.2.5
 
-**NeoForge**
+**Forge**
 
 - Port to 1.18.2